Move schema function to dedicated go file
